improved tests
took into consideration more fringe cases; 96% coverage

+++ src/controllers/user.js
@@ -1,19 +1,20 @@
 const User = require('../models/user');
 
-// newUser function for post user route
+
+// POST /user => createUser
 const createUser = async (req, res, next) => {
     const user = new User({
         email: req.body.email, password: req.body.password
     });
 
     //pw must have 1 lowercase letter, 1 uppercase, 1 number and must be between 8 and 32 characters
-    if (!user.password.match(/^(?=.*\d)(?=.*[A-Z])(?=.*[a-z])([^\s]){8,}$/)) {
+    if (user.password == null || !user.password.match(/^(?=.*\d)(?=.*[A-Z])(?=.*[a-z])([^\s]){8,}$/)) {
         res.status(400).send("failed to create user: insecure password");
         return;
     }
 
     //email address must be valid
-    if (!user.email.match(/^([_a-z0-9]+[\._a-z0-9]*)(\+[a-z0-9]+)?@(([a-z0-9-]+\.)*[a-z]{2,4})$/)) {
+    if (user.email == null || !user.email.match(/^([_a-z0-9]+[\._a-z0-9]*)(\+[a-z0-9]+)?@(([a-z0-9-]+\.)*[a-z]{2,4})$/)) {
         res.status(400).send("failed to create user: invalid email address");
         return;
     }
@@ -33,18 +34,21 @@ const createUser = async (req, res, next) => {
     }
 };
 
+// DELETE /user => deleteUser
 const deleteUser = async (req, res, next) => {
     const user = new User({
         email: req.body.email, password: req.body.password
     });
-    let del_count = (await User.deleteOne({ email: user.email, password: user.password })).deletedCount;
-
-    if (del_count != 0) {
-        if(del_count == null || del_count == undefined)
-            console.log("aaahhhhhhh " + del_count)
-        res.status(200).send("user deleted successfully");
-    } else {
+    if(user.email == null || user.password == null)
         res.status(400).send("failed to delete user");
+    else {
+        let del_count = (await User.deleteOne({ email: user.email, password: user.password })).deletedCount;
+
+        if (del_count != 0) {
+            res.status(200).send("user deleted successfully");
+        } else {
+            res.status(400).send("failed to delete user");
+        }
     }
 }
